+++ packages/zosfiles/src/methods/download/Download.ts
@@ -869,18 +869,11 @@ export class Download {
         const responseLines = [];
 
         if (result.downloaded.length > 0) {
-            if (options.directory) {
-                responseLines.push(
-                    TextUtils.chalk.green(`${result.downloaded.length} data set(s) downloaded successfully to `) +
-                        options.directory,
-                    ...result.downloaded.map(dsname => `    ${dsname}`)
-                );
-            } else {
-                responseLines.push(
-                    TextUtils.chalk.green(`${result.downloaded.length} data set(s) downloaded successfully to ./`),
-                    ...result.downloaded.map(dsname => `    ${dsname}`)
-                );
-            }
+            responseLines.push(
+                TextUtils.chalk.green(`${result.downloaded.length} data set(s) downloaded successfully to `) +
+                    (options.directory ?? "./"),
+                ...result.downloaded.map(dsname => `    ${dsname}`)
+            );
         }
 
         if (result.skippedExisting && result.skippedExisting.length > 0) {